pyproject.toml
[project]
name = "rna-folding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "dvc>=3.66.0",
    "dvc-gdrive>=3.0.1",
    "fire>=0.7.1",
    "hydra-core>=1.3.0",
    "mlflow>=3.8.1",
    "omegaconf>=2.3.0",
    "onnx>=1.20.0",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "pytorch-lightning>=2.6.0",
    "torch>=2.9.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]

[tool.isort]
profile = "black"
